+++ src/lib.rs
@@ -39,6 +39,8 @@ mod egraph;
 mod explain;
 mod extract;
 mod language;
+#[cfg(feature = "lp")]
+mod lp_extract;
 mod machine;
 mod pattern;
 mod rewrite;
@@ -93,6 +95,9 @@ pub use {
     subst::{Subst, Var},
     util::*,
 };
+
+#[cfg(feature = "lp")]
+pub use lp_extract::*;
 
 #[cfg(test)]
 fn init_logger() {